requirements.txt
pytest

tqdm
numpy
pandas
scipy

rdkit-pypi==2022.9.1
